types: extract navigation types to navigation.ts
